Chatbot: Begin Working on AI Component
I fixed an issue with my permissions code and began working on the AI component of the app. For now, I created the different states and got my app to randomly send one of 3-4 equivalent message options as it cycles through each state. However, it does not pay attention to what message the user sends yet.